12_auth/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pyjwt
python-multipart